Add classes and interfaces for Magazine

+++ src/com/example/library/interfaces/BookRepository.java
@@ -7,6 +7,6 @@ import com.example.library.model.Book;
 
 public interface BookRepository {
     List<Book> findAll();
-    Optional<Book> findById(int id);
+    Optional<Book> findById(long id);
     void save(Book book);
 }
